refactor: move owner + related files to internal/muxcore/owner/ (T019)
