+++ internal/muxcore/owner/busy_protocol_test.go
@@ -1,4 +1,4 @@
-package mux
+package owner
 
 import (
 	"testing"
