+++ internal/muxcore/owner/listchanged_inject.go
@@ -1,4 +1,4 @@
-package mux
+package owner
 
 import (
 	"bytes"
