+++ internal/muxcore/owner/listchanged_inject_test.go
@@ -1,4 +1,4 @@
-package mux
+package owner
 
 import (
 	"encoding/json"
