+++ internal/muxcore/owner/owner_serve_test.go
@@ -1,4 +1,4 @@
-package mux
+package owner
 
 import (
 	"context"
