+++ internal/muxcore/owner/progress_reporter.go
@@ -1,4 +1,4 @@
-package mux
+package owner
 
 import (
 	"context"
