+++ internal/muxcore/owner/progress_reporter_test.go
@@ -1,4 +1,4 @@
-package mux
+package owner
 
 import (
 	"bytes"
