+++ internal/muxcore/owner/resilient_client_test.go
@@ -1,4 +1,4 @@
-package mux
+package owner
 
 import (
 	"bufio"
